move to /matrix/chat-message
