+++ src/lib/chat/chat-message.ts
@@ -4,7 +4,6 @@ import { ChatMember } from './types';
 import { denormalize as denormalizeUser } from '../../store/users';
 import { currentUserSelector } from '../../store/authentication/saga';
 import moment from 'moment';
-import { MatrixClient as SDKMatrixClient } from 'matrix-js-sdk';
 
 const DEFAULT_MEDIA_TYPE = 'image';
 
@@ -117,41 +116,6 @@ function extractMessageData(jsonData, isMediaMessage) {
     admin,
     optimisticId,
     rootMessageId,
-  };
-}
-
-async function extractParentMessageData(matrixMessage, sdkMatrixClient: SDKMatrixClient) {
-  const parentMessageData = {
-    parentMessageId: null,
-    parentMessageText: '',
-  };
-
-  const parent = matrixMessage.content['m.relates_to'];
-  if (parent && parent['m.in_reply_to']) {
-    parentMessageData.parentMessageId = parent['m.in_reply_to'].event_id;
-
-    const parentMessage = await sdkMatrixClient.fetchRoomEvent(
-      matrixMessage.room_id,
-      parentMessageData.parentMessageId
-    );
-    if (parentMessage) {
-      parentMessageData.parentMessageText = parentMessage.content.body;
-    }
-  }
-
-  return parentMessageData;
-}
-
-export async function mapMatrixMessage(matrixMessage, sdkMatrixClient: SDKMatrixClient) {
-  return {
-    id: matrixMessage.event_id,
-    message: matrixMessage.content.body,
-    createdAt: matrixMessage.origin_server_ts,
-    updatedAt: null,
-    sender: {},
-    isAdmin: false,
-    ...{ mentionedUsers: [], hidePreview: false, media: null, image: null, admin: {} },
-    ...(await extractParentMessageData(matrixMessage, sdkMatrixClient)),
   };
 }
